VF003-01: polish file structure in frontend with file name improvement.
