+++ src/components/record/AddPetOwner/AddPetOwner.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import AddPetOwnerForm from '../AddPetOwnerForm'
+import AddPetOwnerForm from '../AddPetOwnerForm/AddPetOwnerForm'
 import { useLocation } from 'react-router-dom'
 import { Breadcrumb, Button, Flex } from 'antd'
