--- src/pages/Dashboard.css
@@ -1,14 +1,0 @@
-.main {
-    display: flex;
-    
-    min-height: 100vh;
-    background: #FAFAFB;
-}
-
-
-.mainContent {
-    display: flex;
-    flex-direction: column;
-    margin-left: 2vw;
-}
-
