+++ src/pages/Homepage/Homepage.js
@@ -6,16 +6,16 @@ import {
   useMatch,
   useLocation,
 } from 'react-router-dom'
-import Sidebar from '../../components/publicComponents/Sidebar'
-import Topbar from '../../components/publicComponents/Topbar'
-import HeadnBread from '../../components/HeadnBread'
+import Sidebar from '../../components/publicComponents/Sidebar/Sidebar'
+import Topbar from '../../components/publicComponents/Topbar/Topbar'
+import HeadnBread from '../../components/publicComponents/HeadnBread/HeadAndBread'
 
-import DashboardContent from '../../components/DashboardContent'
-import Appointment from '../../components/Appointment'
-import HealthRecord from '../../components/HealthRecord'
-import AddPet from '../../components/AddPet'
-import Setting from '../../components/Setting'
-import EditProfile from '../../components/EditProfile'
+import DashboardContent from '../../components/Dashboard/Dashboard'
+import Appointment from '../../components/Appointment/Appointment'
+import HealthRecord from '../../components/record/HealthRecord/HealthRecord'
+import AddPet from '../../components/record/AddPet/AddPet'
+import Setting from '../../components/Setting/Setting'
+import EditProfile from '../../components/record/EditProfile/EditProfile'
 
 import './Homepage.css'
 import '../../styles/icon.css'
@@ -30,19 +30,14 @@ const HomePage = () => {
         <HeadnBread />
         <Routes>
           <Route path="" element={<Navigate to="dashboard" />} />
-
           <Route path="dashboard" element={<DashboardContent />} />
-
           <Route path="health-records" element={<HealthRecord />} />
           <Route path="health-records/add-pet" element={<AddPet />} />
-
           <Route path="appointment" element={<Appointment />} />
-
           <Route path="settings" element={<Setting />} />
           <Route path="settings/edit-profile" element={<EditProfile />} />
         </Routes>
       </div>
-
       {/* Add dashboard content here */}
     </div>
   )
